fix: restore test_environment_drift.py's no-growth ceiling, extract promotion tests
Codex P1 finding (PR #1221): the debt.yaml audit for this PR's `compare
--env-matrix` demotion raised eleven `no_growth` ceilings in total. Each was
individually re-verified against its recorded rationale and the actual diff
it represents rather than blanket-reverted.

Ten were genuine, irreducible additions already matching this file's own
"call-site cost lands here, logic lives in a sibling module" convention
(checker.py, diff_versioning.py, cli_compare_receipt.py,
cli_compare_release_matrix.py, cli_helpers_compare.py,
frontends/cli/commands/compare_bundle_facts.py, reporter_markdown.py,
service_compare_pipeline.py, report/render_markdown.py,
report/render_markdown_document.py) -- kept as-is.

tests/test_environment_drift.py's raise (1639 -> 1998, +359 lines) was the
one real instance of the anti-pattern: three classes covering
diff_versioning.promote_baseline_violation_findings
(TestPromoteBaselineViolationFindings, TestWheelRpathNotPortableStaysAtRisk,
TestPromotionRunsBeforeSuppressionRecording) are a self-contained unit
testing one new function, not the drift-report/DT_RELR/hash-style/time64
behavior that module's own docstring scopes it to. Moved verbatim to a new
sibling file, tests/test_diff_versioning_baseline_promotion.py, and
restored the ceiling to the actual measured post-extraction line count
(1750) instead of trimming content to fit.

Verified: mypy abicheck/ clean, ruff check clean, python
scripts/check_architecture.py 0 errors, python scripts/check_ai_readiness.py
clean (pre-existing unrelated adr-status-sync errors excepted), both test
files pass (169 passed).
