+++ architecture/debt.yaml
@@ -2093,13 +2093,13 @@
     },
     {
       "path": "tests/test_environment_drift.py",
-      "baseline_lines": 1998,
+      "baseline_lines": 1750,
       "target": "tests/unit-or-contract ownership matching production migration",
       "rule": "no_growth",
       "disposition": "migrate",
       "category": "legacy_oversized_test_module",
       "owner": "abicheck maintainers",
-      "rationale": "Existing test module predates ADR-061 and will split by behavior or contract axis with its production migration. Raised 1639 -> 1736 (compare --env-matrix demotion, ADR-068 D5): this is the module's own canonical test home for the declared-deployment-matrix contract (its own module docstring), so the config-driven CLI coverage the demotion requires -- `.abicheck.yml` `deployment:` producing the identical verdict/exit-code behavior the retired `--env-matrix FILE` used to, the retired flag itself exiting 64, and the config-only key now applying (rather than being rejected) across a directory/package release fan-out -- lands here rather than a new sibling file. `TestPlatformBaselineFloorCliEndToEnd`'s two existing CLI tests were ported from `--env-matrix FILE` to `--config .abicheck.yml`; one new test in that class and two new tests replacing the old release-rejection test cover the retired-flag and fan-out-applies cases; `TestLoadEnvMatrix`'s `CompareRequest.env_matrix_path`-existence test was rewritten for the field's replacement (`CompareRequest.env_matrix`, an already-resolved value, not a path to validate). No smaller diff: each test exercises a real, distinct behavior this demotion changed. Raised 1736 -> 1872 (Codex review, PR #1221, no-baseline env-matrix audit finding 1): this module is also the canonical home for `diff_versioning.py`'s declared-runtime-floor/wheel-packaging checks (see its existing `TestPlatformBaselineFloorRaised`/`TestMusllinuxGlibcDependency` classes), so the new `TestPromoteBaselineViolationFindings` class -- a property-style class stating the new `promote_baseline_violation_findings()` promotion invariant generally over every promoted kind, a negative control over the unpromoted/unrelated kinds, and a two-sided/no-baseline parity test -- lands here rather than a new sibling file. Two pre-existing tests in this same module (`TestPlatformBaselineFloorRaised`/`TestPlatformBaselineFloorCliEndToEnd`) were also updated in place for the corrected BREAKING verdict, at no additional line cost beyond the rename/assertion churn. Raised 1872 -> 1993 (Codex review, PR #1221, follow-up findings 1/3): `TestPromoteBaselineViolationFindings` shrank to two promoted kinds (`WHEEL_RPATH_NOT_PORTABLE` is no longer promoted -- see `diff_versioning.py`'s own updated rationale) and its own two-sided/no-baseline parity test dropped the now-inapplicable macOS/WHEEL_RPATH cases; a new sibling class, `TestWheelRpathNotPortableStaysAtRisk`, states the complementary invariant this same review split out -- that kind must stay at its catalog-default RISK on both paths -- so the promoted-vs-not-promoted contract is covered by two focused classes rather than one class asserting two different things about two different subsets. A second new class, `TestPromotionRunsBeforeSuppressionRecording`, covers finding 3: a suppression rule matching a genuine `PLATFORM_BASELINE_FLOOR_RAISED` floor violation must still record it in `suppressed_changes` at the correctly-promoted BREAKING verdict, not the unpromoted RISK default, since promotion now runs before `_filter_suppressed_changes` disposes of the match (AGENTS.md's \"Record before disposing\") -- checked against an unsuppressed control asserting the identical violation's own BREAKING verdict, so the test cannot pass by coincidentally matching a stale expectation on either side. This module remains the canonical home for both invariants per this entry's own prior raise. Raised 1993 -> 1998 (Codex review, PR #1221, round-6 follow-up finding 1): already at the AI-readiness 2000-line hard cap, so the new `WHEEL_ARCH`/`MUSLLINUX`/`WHEEL_CONTEXT` non-string-value tests landed in a new sibling file, `tests/test_environment_matrix_floor_types.py`, instead of growing this one further; the five-line growth here is a `TestEnvironmentMatrixRuntimeFloors`-scoped comment pointing there.",
+      "rationale": "Existing test module predates ADR-061 and will split by behavior or contract axis with its production migration. Raised 1639 -> 1750 (compare --env-matrix demotion, ADR-068 D5): this is the module's own canonical test home for the declared-deployment-matrix contract (its own module docstring), so the config-driven CLI coverage the demotion requires -- `.abicheck.yml` `deployment:` producing the identical verdict/exit-code behavior the retired `--env-matrix FILE` used to, the retired flag itself exiting 64, and the config-only key now applying (rather than being rejected) across a directory/package release fan-out -- lands here rather than a new sibling file. `TestPlatformBaselineFloorCliEndToEnd`'s two existing CLI tests were ported from `--env-matrix FILE` to `--config .abicheck.yml`; one new test in that class and two new tests replacing the old release-rejection test cover the retired-flag and fan-out-applies cases; `TestLoadEnvMatrix`'s `CompareRequest.env_matrix_path`-existence test was rewritten for the field's replacement (`CompareRequest.env_matrix`, an already-resolved value, not a path to validate). No smaller diff: each test exercises a real, distinct behavior this demotion changed. This total also folds in `TestPlatformBaselineFloorRaised`'s pre-existing test being updated in place for `diff_versioning.py`'s own unrelated `promote_baseline_violation_findings` BREAKING-verdict correction (rename/assertion churn only, no added lines) and a five-line `TestEnvironmentMatrixRuntimeFloors`-scoped comment pointing at `tests/test_environment_matrix_floor_types.py` for the `WHEEL_ARCH`/`MUSLLINUX`/`WHEEL_CONTEXT` non-string-value tests that landed there instead, since this module was already at the AI-readiness 2000-line hard cap. Architecture-debt no-growth review (Codex P1 finding on this same PR, fresh audit of every ceiling this PR raised): the three classes this PR had added for `promote_baseline_violation_findings` itself (`TestPromoteBaselineViolationFindings`, `TestWheelRpathNotPortableStaysAtRisk`, `TestPromotionRunsBeforeSuppressionRecording` -- previously raising this baseline to 1998, +359 over the pre-PR 1639) were a self-contained unit testing one new `diff_versioning.py` function and its two immediate corrections, not the drift-report/DT_RELR/hash-style/time64 behavior this module's own docstring scopes it to -- a genuinely separable axis, unlike the demotion coverage above which this module's docstring already claims as its own canonical home. Extracted verbatim to the new sibling `tests/test_diff_versioning_baseline_promotion.py` (own `no_growth` baseline below) rather than kept here or trimmed, per this file's own \"move responsibility out, never trim to fit\" convention -- restoring this entry to the actual measured post-extraction line count.",
       "review_by": "2026-11-30"
     },
     {
